+++ lib/vanagon/utilities.rb
@@ -34,9 +34,9 @@ class Vanagon
 
     # If Digest::const_get fails, it'll raise a LoadError when it tries to
     # pull in the subclass `type`. We catch that error, and fail instead.
-     rescue LoadError
-       fail "Don't know how to produce a sum of type: '#{type}' for '#{file}'"
-     end
+    rescue LoadError
+      fail "Don't know how to produce a sum of type: '#{type}' for '#{file}'"
+    end
 
     # Simple wrapper around Net::HTTP. Will make a request of the given type to
     # the given url and return the body as parsed by JSON.
@@ -112,9 +112,12 @@ class Vanagon
     # @return [String, false] Returns either the full path to the command or false if the command cannot be found
     # @raise [RuntimeError] If the command is required and cannot be found
     def find_program_on_path(command, required = true)
+      extensions = ENV['PATHEXT'] ? ENV['PATHEXT'].split(';') : ['']
       ENV['PATH'].split(File::PATH_SEPARATOR).each do |path_elem|
-        location = File.join(path_elem, command)
-        return location if FileTest.executable?(location)
+        extensions.each do |ext|
+          location = File.join(path_elem, "#{command}#{ext}")
+          return location if FileTest.executable?(location)
+        end
       end
 
       if required