+++ spec/lib/vanagon/project_spec.rb
@@ -1,10 +1,17 @@
 require 'vanagon/project'
 require 'vanagon/driver'
 require 'vanagon/errors'
+require 'fakefs/spec_helpers'
 
 describe 'Vanagon::Project' do
   let(:component) { double(Vanagon::Component) }
   let(:configdir) { '/a/b/c' }
+  let(:platform) do
+    OpenStruct.new(:settings => {})
+  end
+  let(:upstream_platform) do
+    OpenStruct.new(:settings => {})
+  end
 
   let(:project_block) {
     "project 'test-fixture' do |proj|
@@ -45,15 +52,17 @@ describe 'Vanagon::Project' do
     end"
   }
 
-  let (:dummy_platform_sysv) {
+  let (:dummy_platform_settings) {
     plat = Vanagon::Platform::DSL.new('debian-6-i386')
     plat.instance_eval("platform 'debian-6-i386' do |plat|
                        plat.servicetype 'sysv'
                        plat.servicedir '/etc/init.d'
                        plat.defaultdir '/etc/default'
+                       settings[:platform_test] = 'debian'
                     end")
     plat._platform
   }
+
 
   describe '#vendor=' do
     dummy_platform = Vanagon::Platform.new('el-7-x86_64')
@@ -112,7 +121,7 @@ describe 'Vanagon::Project' do
     it 'returns only the highest level directories' do
       test_sets.each do |set|
         expect(component).to receive(:directories).and_return([])
-        proj = Vanagon::Project::DSL.new('test-fixture', {}, [])
+        proj = Vanagon::Project::DSL.new('test-fixture', configdir, platform, [])
         proj.instance_eval(project_block)
         set[:directories].each {|dir| proj.directory dir }
         expect(proj._project.get_root_directories.sort).to eq(set[:results].sort)
@@ -129,34 +138,55 @@ describe 'Vanagon::Project' do
       expect(git_source).to receive(:fetch).and_return(true)
 
       # stubs for the upstream project
-      upstream_proj = Vanagon::Project::DSL.new('upstream-test', {}, [])
+      upstream_proj = Vanagon::Project::DSL.new('upstream-test', configdir, upstream_platform, [])
       upstream_proj.instance_eval(upstream_project_block)
       expect(Vanagon::Project).to receive(:load_project).and_return(upstream_proj._project)
+      expect(Vanagon::Platform).to receive(:load_platform).and_return(upstream_platform)
+
+      class Vanagon
+        class Project
+          BUILD_TIME = '2017-07-10T13:34:25-07:00'
+          VANAGON_VERSION = '0.0.0-rspec'
+        end
+      end
+
     end
 
     it 'loads upstream settings' do
-      inheriting_proj = Vanagon::Project::DSL.new('inheritance-test', {}, [])
+      inheriting_proj = Vanagon::Project::DSL.new('inheritance-test', configdir, platform, [])
       inheriting_proj.instance_eval(inheriting_project_block)
       expect(inheriting_proj._project.settings[:test]).to eq('upstream-test')
     end
 
     it 'overrides duplicate settings from before the load' do
-      inheriting_proj = Vanagon::Project::DSL.new('inheritance-test', {}, [])
+      inheriting_proj = Vanagon::Project::DSL.new('inheritance-test', configdir, platform, [])
       inheriting_proj.instance_eval(preset_inheriting_project_block)
       expect(inheriting_proj._project.settings[:test]).to eq('upstream-test')
     end
 
     it 'lets you override settings after the load' do
-      inheriting_proj = Vanagon::Project::DSL.new('inheritance-test', {}, [])
+      inheriting_proj = Vanagon::Project::DSL.new('inheritance-test', configdir, platform, [])
       inheriting_proj.instance_eval(postset_inheriting_project_block)
       expect(inheriting_proj._project.settings[:test]).to eq('inheritance-test')
     end
 
     it 'merges settings' do
-      inheriting_proj = Vanagon::Project::DSL.new('inheritance-test', {}, [])
+      inheriting_proj = Vanagon::Project::DSL.new('inheritance-test', configdir, platform, [])
       inheriting_proj.instance_eval(inheriting_project_block_with_settings)
       expect(inheriting_proj._project.settings[:test]).to eq('upstream-test')
       expect(inheriting_proj._project.settings[:merged]).to eq('yup')
+    end
+  end
+
+  describe 'platform settings' do
+    before do
+      allow(Vanagon::Component).to receive(:load_component).with('some-component', any_args).and_return(component)
+    end
+
+    it 'loads settings set in platforms' do
+      settings_proj = Vanagon::Project::DSL.new('settings-test', configdir, dummy_platform_settings, [])
+      settings_proj.instance_eval(project_block)
+      expect(settings_proj._project.settings[:platform_test]).to eq('debian')
     end
   end
 
@@ -236,7 +266,7 @@ describe 'Vanagon::Project' do
     # All of the following tests should be run with one project level
     # component that isn't included in the build_deps of another component
     before(:each) do
-      @proj = Vanagon::Project.new('test-fixture-with-comps', {})
+      @proj = Vanagon::Project.new('test-fixture-with-comps', platform)
       @not_included_comp = Vanagon::Component.new('test-fixture-not-included', {}, {})
       @proj.components << @not_included_comp
     end
@@ -305,7 +335,7 @@ describe 'Vanagon::Project' do
 
   describe '#get_preinstall_actions' do
     it "Collects the preinstall actions for the specified package state" do
-      proj = Vanagon::Project.new('action-test', {})
+      proj = Vanagon::Project.new('action-test', platform)
       proj.get_preinstall_actions('upgrade')
       proj.get_preinstall_actions('install')
       expect(proj.get_preinstall_actions('install')).to be_instance_of(String)
@@ -314,39 +344,39 @@ describe 'Vanagon::Project' do
 
   describe '#get_trigger_scripts' do
     it "Collects the install triggers for the project for the specified packing state" do
-      proj = Vanagon::Project.new('action-test', {})
+      proj = Vanagon::Project.new('action-test', platform)
       expect(proj.get_trigger_scripts('install')).to eq({})
       expect(proj.get_trigger_scripts('upgrade')).to be_instance_of(Hash)
     end
     it 'fails with empty install trigger action' do
-      proj = Vanagon::Project.new('action-test', {})
+      proj = Vanagon::Project.new('action-test', platform)
       expect { proj.get_trigger_scripts([]) }.to raise_error(Vanagon::Error)
     end
     it 'fails with incorrect install trigger action' do
-      proj = Vanagon::Project.new('action-test', {})
+      proj = Vanagon::Project.new('action-test', platform)
       expect { proj.get_trigger_scripts('foo') }.to raise_error(Vanagon::Error)
     end
   end
 
   describe '#get_interest_triggers' do
     it "Collects the interest triggers for the project for the specified packaging state" do
-      proj = Vanagon::Project.new('action-test', {})
+      proj = Vanagon::Project.new('action-test', platform)
       expect(proj.get_interest_triggers('install')).to eq([])
       expect(proj.get_interest_triggers('upgrade')).to be_instance_of(Array)
     end
     it 'fails with empty interest trigger action' do
-      proj = Vanagon::Project.new('action-test', {})
+      proj = Vanagon::Project.new('action-test', platform)
       expect { proj.get_interest_triggers([]) }.to raise_error(Vanagon::Error)
     end
     it 'fails with incorrect interest trigger action' do
-      proj = Vanagon::Project.new('action-test', {})
+      proj = Vanagon::Project.new('action-test', platform)
       expect { proj.get_interest_triggers('foo') }.to raise_error(Vanagon::Error)
     end
   end
 
   describe '#get_activate_triggers' do
     it "Collects the activate triggers for the project for the specified packaging state" do
-      proj = Vanagon::Project.new('action-test', {})
+      proj = Vanagon::Project.new('action-test', platform)
       expect(proj.get_activate_triggers()).to be_instance_of(Array)
       expect(proj.get_activate_triggers()).to be_instance_of(Array)
     end
@@ -354,7 +384,7 @@ describe 'Vanagon::Project' do
 
   describe '#generate_dependencies_info' do
     before(:each) do
-      @proj = Vanagon::Project.new('test-project', {})
+      @proj = Vanagon::Project.new('test-project', platform)
     end
 
     it "returns a hash of components and their versions" do
@@ -385,7 +415,7 @@ describe 'Vanagon::Project' do
         end
       end
 
-      @proj = Vanagon::Project.new('test-project', {})
+      @proj = Vanagon::Project.new('test-project', platform)
     end
 
     it 'should generate a hash with the expected build metadata' do
@@ -412,6 +442,50 @@ describe 'Vanagon::Project' do
       @proj.build_manifest_json(true)
     end
   end
+
+  describe '#save_manifest_json' do
+    include FakeFS::SpecHelpers
+    let(:platform_name) { 'el-7-x86_64' }
+    let(:platform) { Vanagon::Platform.new(platform_name) }
+    before(:each) do
+      class Vanagon
+        class Project
+          BUILD_TIME = '2018-07-10T13:34:25-07:00'
+          VANAGON_VERSION = '0.0.0-rspec'
+        end
+      end
+
+      @proj = Vanagon::Project.new('test-project', platform)
+    end
+
+    it 'should generate a file with the expected build metadata' do
+      correct_sample_metadata = {
+        'packaging_type' => { 'vanagon' => '0.0.0-rspec' },
+        'version' => '123abcde',
+        'components' => { 'test-component-10' => { 'version' => '1.2.3' } },
+        'build_time' => '2018-07-10T13:34:25-07:00',
+      }
+      bad_sample_metadata = {
+        'BAD KEY' => 'BAD VALUE'
+      }
+      comp1 = Vanagon::Component.new('test-component-10', {}, {})
+      comp1.version = '1.2.3'
+      @proj.components << comp1
+      @proj.version = '123abcde'
+      FakeFS do
+        @proj.save_manifest_json(platform)
+
+        old_style_metadata = JSON.parse(File.read('ext/build_metadata.json'))
+        expect(old_style_metadata).to eq(correct_sample_metadata)
+
+        metadata_with_project_and_platform = JSON.parse(
+          File.read("ext/build_metadata.test-project.#{platform_name}.json"))
+        expect(metadata_with_project_and_platform).to eq(correct_sample_metadata)
+        expect(metadata_with_project_and_platform).not_to eq(bad_sample_metadata)
+      end
+    end
+  end
+
 
   describe '#publish_yaml_settings' do
     let(:platform_name) { 'aix-7.2-ppc' }
@@ -465,7 +539,7 @@ describe 'Vanagon::Project' do
     it "builds packages by default" do
       platform = Vanagon::Platform::DSL.new('el-7-x86_64')
       platform.instance_eval("platform 'el-7-x86_6' do |plat| end")
-      proj = Vanagon::Project::DSL.new('test-fixture', platform._platform, [])
+      proj = Vanagon::Project::DSL.new('test-fixture', configdir, platform._platform, [])
       expect(platform._platform).to receive(:generate_package) { ["# making a package"] }
       expect(proj._project.generate_package).to eq(["# making a package"])
     end
@@ -473,7 +547,7 @@ describe 'Vanagon::Project' do
     it "builds packages and archives if configured for both" do
       platform = Vanagon::Platform::DSL.new('el-7-x86_64')
       platform.instance_eval("platform 'el-7-x86_6' do |plat| end")
-      proj = Vanagon::Project::DSL.new('test-fixture', platform._platform, [])
+      proj = Vanagon::Project::DSL.new('test-fixture', configdir, platform._platform, [])
       proj.generate_archives(true)
       expect(platform._platform).to receive(:generate_package) { ["# making a package"] }
       expect(platform._platform).to receive(:generate_compiled_archive) { ["# making an archive"] }
@@ -483,7 +557,7 @@ describe 'Vanagon::Project' do
     it "can build only archives" do
       platform = Vanagon::Platform::DSL.new('el-7-x86_64')
       platform.instance_eval("platform 'el-7-x86_6' do |plat| end")
-      proj = Vanagon::Project::DSL.new('test-fixture', platform._platform, [])
+      proj = Vanagon::Project::DSL.new('test-fixture', configdir, platform._platform, [])
       proj.generate_archives(true)
       proj.generate_packages(false)
       expect(platform._platform).to receive(:generate_compiled_archive) { ["# making an archive"] }
@@ -493,9 +567,24 @@ describe 'Vanagon::Project' do
     it "builds nothing if that's what you really want" do
       platform = Vanagon::Platform::DSL.new('el-7-x86_64')
       platform.instance_eval("platform 'el-7-x86_6' do |plat| end")
-      proj = Vanagon::Project::DSL.new('test-fixture', platform._platform, [])
+      proj = Vanagon::Project::DSL.new('test-fixture', configdir, platform._platform, [])
       proj.generate_packages(false)
       expect(proj._project.generate_package).to eq([])
+    end
+  end
+
+  describe '#get_rpm_ghost_files' do
+    it 'returns an empty array when there are no ghost files' do
+      proj = Vanagon::Project.new('test-ghost', platform)
+      expect(proj.get_rpm_ghost_files).to eq([])
+    end
+
+    it 'returns ghost files when some are set' do
+      proj = Vanagon::Project.new('test-ghosts', platform)
+      comp = Vanagon::Component.new('ghosts', {}, {})
+      comp.add_rpm_ghost_file('ghost')
+      proj.components << comp
+      expect(proj.get_rpm_ghost_files).to eq(['ghost'])
     end
   end
 end